Select branch as a part of clone
One can specify the branch to track in the same 'git' command as the
clone. This can make the operation a little quicker if the desired
branch is significantly different from the default (typically
'master') branch.

It will also facilitate the desired branch being checked out in future
commits that implement shallow clones.