Allow gitpullers to take and act on a depth parameter
If provided to the GitPuller, a 'depth' parameter will be passed along
to git, resulting in a shallow clone. Shallow clones with a depth of
'N' only include history going back 'N' commits. This can be helpful
in reducing the size of the cloned repository, and in the amount of
time it takes to clone that repository, especially when large files
were added and later removed/edited.

As a special-case, a non-positive value for depth (eg, 0 or negative)
can be specified to explicitly request all history.

For more information, see:
https://git-scm.com/docs/git-clone#git-clone---depthltdepthgt