Moved the two constants and REPO_PARENT_DIR out of __init__.py
The REPO_PARENT_DIR never belonged there. I now resolve it in
Synchandler as we always did and pass into the handle_files function
as part of the args; it will be used in the plugin_helper.py as the path
to the files.

The other two TEMP_DOWNLOAD_REPO_SIR and CACHED_ORIGIN_NON_GIT_REPO
are now declared and initialized in plugin_helper.py where they are
used.