When resetting deleted files, take from local copy
Fixes issues #254 and #234.
When a deleted file is reset, this checks it out from the local copy
of the branch, rather than from the remote copy. The file might not
exist in the remote branch anymore, causing #234 when running
`git checkout`. If the file was conflicted before, we will run into
issue #254.

If the remote file has changed, it will be merged regularly afterwards.